+++ indexer_test.go
@@ -7,12 +7,13 @@ import (
 	"io/ioutil"
 	"log"
 	"net/http"
+	"net/http/httptest"
 	"os"
 	"testing"
 	"time"
 
 	_ "github.com/lib/pq"
-	"github.com/olivere/elastic"
+	"github.com/olivere/elastic/v7"
 	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -48,9 +49,9 @@ func setup(t *testing.T) (*sql.DB, *elastic.Client) {
 func assertQuery(t *testing.T, client *elastic.Client, index string, query elastic.Query, hits []int64) {
 	results, err := client.Search().Index(index).Query(query).Sort("id", true).Pretty(true).Do(context.Background())
 	assert.NoError(t, err)
-	assert.Equal(t, int64(len(hits)), results.Hits.TotalHits)
+	assert.Equal(t, int64(len(hits)), results.Hits.TotalHits.Value)
 
-	if int64(len(hits)) == results.Hits.TotalHits {
+	if int64(len(hits)) == results.Hits.TotalHits.Value {
 		for i, hit := range results.Hits.Hits {
 			assert.Equal(t, fmt.Sprintf("%d", hits[i]), hit.Id)
 		}
@@ -93,6 +94,9 @@ func TestIndexing(t *testing.T) {
 
 	assertQuery(t, client, physicalName, elastic.NewMatchQuery("is_blocked", "true"), []int64{3})
 	assertQuery(t, client, physicalName, elastic.NewMatchQuery("is_stopped", "true"), []int64{2})
+
+	assertQuery(t, client, physicalName, elastic.NewMatchQuery("status", "B"), []int64{3})
+	assertQuery(t, client, physicalName, elastic.NewMatchQuery("status", "S"), []int64{2})
 
 	assertQuery(t, client, physicalName, elastic.NewMatchQuery("org_id", "1"), []int64{1, 2, 3, 4})
 
@@ -198,12 +202,12 @@ func TestIndexing(t *testing.T) {
 	// phrase matches all
 	query = elastic.NewNestedQuery("fields", elastic.NewBoolQuery().Must(
 		elastic.NewMatchQuery("fields.field", "fcab2439-861c-4832-aa54-0c97f38f24ab"),
-		elastic.NewMatchPhraseQuery("fields.district", "King Côunty")))
+		elastic.NewMatchPhraseQuery("fields.district", "King-Côunty")))
 	assertQuery(t, client, physicalName, query, []int64{7})
 
 	query = elastic.NewNestedQuery("fields", elastic.NewBoolQuery().Must(
 		elastic.NewMatchQuery("fields.field", "fcab2439-861c-4832-aa54-0c97f38f24ab"),
-		elastic.NewMatchQuery("fields.district_keyword", "King Côunty")))
+		elastic.NewMatchQuery("fields.district_keyword", "King-Côunty")))
 	assertQuery(t, client, physicalName, query, []int64{7})
 
 	// ward query
@@ -294,4 +298,87 @@ func TestIndexing(t *testing.T) {
 	// 3 is no longer in our group
 	assertQuery(t, client, indexName, elastic.NewMatchQuery("groups", "529bac39-550a-4d6f-817c-1833f3449007"), []int64{1})
 
+}
+func TestRetryServer(t *testing.T) {
+	responseCounter := 0
+	responses := []func(w http.ResponseWriter, r *http.Request){
+		func(w http.ResponseWriter, r *http.Request) {
+			w.Header().Set("Content-Length", "5")
+		},
+		func(w http.ResponseWriter, r *http.Request) {
+			w.Header().Set("Content-Length", "1")
+		},
+		func(w http.ResponseWriter, r *http.Request) {
+			w.Header().Set("Content-Length", "1")
+		},
+		func(w http.ResponseWriter, r *http.Request) {
+			resp := `{
+				"took": 1,
+				"timed_out": false,
+				"_shards": {
+				  "total": 2,
+				  "successful": 2,
+				  "skipped": 0,
+				  "failed": 0
+				},
+				"hits": {
+				  "total": 1,
+				  "max_score": null,
+				  "hits": [
+					{
+					  "_index": "rp_elastic_test_2020_08_14_1",
+					  "_type": "_doc",
+					  "_id": "1",
+					  "_score": null,
+					  "_routing": "1",
+					  "_source": {
+						"id": 1,
+						"org_id": 1,
+						"uuid": "c7a2dd87-a80e-420b-8431-ca48d422e924",
+						"name": null,
+						"language": "eng",
+						"is_stopped": false,
+						"is_blocked": false,
+						"is_active": true,
+						"created_on": "2017-11-10T16:11:59.890662-05:00",
+						"modified_on": "2017-11-10T16:11:59.890662-05:00",
+						"last_seen_on": "2020-08-04T21:11:00-04:00",
+						"modified_on_mu": 1.510348319890662e15,
+						"urns": [
+						  {
+							"scheme": "tel",
+							"path": "+12067791111"
+						  },
+						  {
+							"scheme": "tel",
+							"path": "+12067792222"
+						  }
+						],
+						"fields": [
+						  {
+							"text": "the rock",
+							"field": "17103bb1-1b48-4b70-92f7-1f6b73bd3488"
+						  }
+						],
+						"groups": [
+						  "4ea0f313-2f62-4e57-bdf0-232b5191dd57",
+						  "529bac39-550a-4d6f-817c-1833f3449007"
+						]
+					  },
+					  "sort": [1]
+					}
+				  ]
+				}
+			  }`
+
+			w.Write([]byte(resp))
+		},
+	}
+	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		responses[responseCounter](w, r)
+		responseCounter++
+	}))
+	defer ts.Close()
+	FindPhysicalIndexes(ts.URL, "rp_elastic_test")
+	require.Equal(t, responseCounter, 4)
 }